feat: improved module init
I was having troubles when having NgxsModule.forFeature in the main app.

It seems that the provider store token was being injected multiple times which caused feature's to be loaded twice.

By separating out the feature states into it's own token it seems to be working better.

I Also added some units tests to test out the different setups we can encounter.