fix(store): improve type signatures of store select methods
- remove overload taking string or any and returning Observable<any>
- add overload taking a class definition and returning Obervable<any>

The motivation of this change is to fix an issue where the result
of calling store.select(MyState.someSelector) is not correctly
inferring the type of the result observable.

As far as I could tell there was no use case for passing a string
to the function, so I am unsure why string is accepted previously.