feat(store): strongly type ofAction* methods
When using the actions stream and the ofAction* operators, restrict the
parameters of these functions to valid action types and the resulting stream
to the union type of the set of actions supplied to the operators.

Eg, before:

```
this.store.actions.pipe(
  ofActionSuccessful(MyAction1, MyAction2),
  tap((a/* : any */) => { });
```

After:

```
this.store.actions.pipe(
  ofActionSuccessful(MyAction1, MyAction2),
  tap((a/* : MyAction1 | MyAction2 */) => { });
```

This change may identify broken code in downstream applications.

closes #1807