refactor: simplify integration app and run Cypress SSR tests (#2017)
This commit removes `@Select` decorator usages from the integration app.
This also updated the Cypress configuration to match `ssr` folder if the
`SSR` environment variable is provided. This is done because the `--integrationFolder`
argument is not available anymore and we should be using `e2e.specPattern` property.
The `integration/project.json` has been updated to run `serve` target with a development
configuration so recompilations are faster by default; its `build` target has been replaced.
The HMR plugin has been removed from the integration app because it's not being maintained
anymore and brings some issues.
with `build-app`.