fix(storage-plugin): access local and session storages globals only in browser
This commit modifies the factory functions for the `LOCAL_STORAGE_ENGINE`
and `SESSION_STORAGE_ENGINE` tokens. Previously, these functions resolved
to `localStorage` and `sessionStorage`, which caused an error when executing
the code in a server-side environment where localStorage is not defined.
To address this, we have implemented a check to determine if the code is
running in a browser environment before retrieving the `localStorage`.