fix(devtools-plugin): do not re-enter Angular zone when resetting state
This commit removes the re-entering of the Angular zone from the devtools plugin.
It is not required to run `store.reset` inside the `ngZone.run`. The `store.reset`
will trigger a state update, but we already have a `leaveNgxs` operator that pipes
selectors. Using `ngZone.run` would result in redundant change detection.