fix(storage-plugin): expose storage plugin internals privately through subpackage (#2052)
This commit exposes some of the storage plugin internals through a subpackage, following a
similar approach to how router plugin internals are exposed. These internals are necessary to
expose if someone wishes to extend the storage engine's behavior. For example, to check for the
default state key `@@STATE` or to use helper functions like extracting string keys.