fix(store): show error when state initialization order is invalid (#2066)
This commit adds a `console.error` when the `UpdateState` is dispatched before
the `InitState`. This typically indicates that the state initialization order is
invalid and must be updated to ensure that states are initialized in the correct
order. The incorrect order may prevent `ngxsOnInit` from being called on feature
states or lead to other unpredictable errors.