feat(storage-plugin): require explicit options when providing storage plugin
This commit updates the implementation of the register signature for the storage plugin.
It now always requires `options` to be provided. We want to ensure clarity regarding the
states that we're going to serialize. When no `options` are provided, we implicitly provide
the `@@STATE` key, which defines that all states should be serialized. This implementation
doesn't compromise extendability because we implicitly rely on the internal mechanism that
transforms options. We should always require options to be provided and require developers
to specify serializable states explicitly.

This would allow us to have feature states because the feature registration mechanism would
also know if the user explicitly specified all states to be serialized at the root level or not.

**Note:** This change is breaking. However, it's time to make breaking changes to move forward,
as the current implementation is stale and acts as a blocker.