feat(storage-plugin): allow providing feature states (#1994)
This commit introduces the ability to specify states for persistence at the feature level.
Previously, you could only specify a list of states for persistence when providing the storage
plugin at the root level. This feature is only available with the standalone API, as Angular has
an `ENVIRONMENT_INITIALIZER` feature that allows us to add these states at the feature level.

The API is simple and straightforward. It introduces the `withStorageFeature` function, which
can be called along with `provideStates`. It requires a list of states to be provided. If all
states are already being persisted because the developer specified `keys: *` at the root level,
this won't do anything.