fix(store): `ctx.getState` should return `DeepReadOnly<>`
This commit modifies the return value of `StateContext.getState` from `<T>` to `ReadonlyDeep<T>`.
The `ReadonlyDeep` type is provided by the `type-fest` package, which is now included in the
dependencies of `@ngxs/store` and will be installed automatically with it.

This change is necessary because `getState` should return an immutable object, and users should
not have any reason to directly modify it.

Please note that this change is breaking and is part of the v4 release preparation.