fix(store): `dispatch` return observable should be `<void>`
This commit updates all `dispatch` signatures to return `Observable<void>`. Initially, the
signature was intended to be `void`, but historically, it was returning `Observable<any>`,
where `any` represents the state snapshot.

This is a breaking change. We've been delaying this fix for years, but now all the necessary
changes should finally be made to allow us to progress with the NGXS code itself. It's especially
important to clean up old planned todos as well.