fix(store): enable throwing errors from selectors by default
This commit updates the `suppressErrors` configuration property to default to falsy.
Consequently, if the selector function throws a type error (such as when accessing a
property unsafely), the error will now be thrown.

`selectSnapshot` will throw synchronously, while errors from `select` will be thrown
asynchronously because the RxJS error reporter operates within a `setTimeout`.

We log errors in development mode to assist developers in identifying selectors that have thrown.

This change is breaking and is part of the v4 release preparation.