fix(store): `setState` and `patchState` should both return `<void>` (#2114)
This commit modifies the signatures of `setState` and `patchState` to return `<void>`.
Previously, they were returning `<T>`, which was historically a mistake. This change also
impacts the `Store.reset` method since it utilizes `setState`, resulting in `Store.reset`
now also returning `<void>`.

`setState` is designed to update the state, and as such, the method itself should not generate
a return value. Users should utilize `getState` to retrieve the latest value accordingly.