fix(store): do not inject container state by default (#2117)
This commit updates the `injectContainerState` configuration property to default to false.
This change will affect the behavior of `@Selector()`s, as they will no longer receive the
container state as the first argument if there are any selectors in the list.

The current behavior is as follows:

```ts
@State()
class NotificationsState {
  @Selector()
  static getNotifications(state: NotificationsStateModel) {
    return state.notifications;
  }

  @Selector([NotificationsState.getNotifications])
  static getUnreadTotal(state: NotificationsStateModel, notifications: Notification[]) {
    return notifications.reduce(() => ..., 0);
  }
}
```

You might notice that the `getUnreadTotal` selector will have the notifications state injected
as the first argument before accessing the notifications list.

This update has been made to prevent injecting the container state as the first argument if
there are any selectors provided in the `@Selector([...])` decorator. However, this change does
not affect the `@Selector()` decorator with no arguments.