fix(store): remove `defaultsState` config property
This commit removes the `defaultsState` configuration property, which was likely a historical
mistake and was used for testing purposes to initialize the state stream before NGXS was
initialized.

Since this property was not documented, it should not have been used by external consumers.
If anyone is affected by this change, they should simply remove it.