fix(store): use `computed` in `selectSignal`
Note: it is necessary to revert these changes and cease experimenting
with decoupling signal changes.

Previously, an issue was reported where calling `dispatch` from within an
effect caused an error indicating that signal writes are not allowed. We attempted
to resolve this by decoupling state signal changes through debouncing them, ensuring
that consumers always received signal updates asynchronously. However, we should no
longer pursue this approach because it fixes initial issues while introducing new ones.
It is the responsibility of developers to ensure a valid reactive context when updating
signals. The `dispatch` function can be wrapped in `untracked`, as it is common practice
to disallow updates to ANY signal within effects, not just NGXS state.

Debouncing signals disrupts zoneless server-side rendering because Angular is unaware
of tasks scheduled by the `debounceTime` operator. Consequently, content is serialized
before the `debounceTime` timer fires internally, which updates the signal and allows
the view to reflect the updated value.