fix(store): disallow nullable actions in `dispatch()`
In this commit, we update the `dispatch()` implementation to account for nullable parameters.
For example, `dispatch(null)` is invalid because `null` is not a valid action. Changing the type
from `any` to `NonNullable<unknown>` is a breaking change. To make this transition more
user-friendly, we include a check, guarded by `ngDevMode`, that throws an error immediately
if `actions` is `null` or `undefined`. This change will only trigger runtime errors in
development mode, ensuring that no production applications are affected.