fix(logger-plugin): run `filter` function in injection context (#2236)
In this commit, we wrap the `options.filter` function for the logger plugin to run within
the injection context, allowing the retrieval of any injectees that may have properties
required for filtering. For example, the user may want to skip logging actions until the
app's bootstrap is complete.

Additionally, the `isIE` check has been removed from the log writer, as Internet Explorer
is no longer supported by Angular.