refactor(store): use `Object.is` as default equality check
In this commit, we default to using `Object.is` as the equality check function
(this function is also used by other frameworks). `Object.is` supports checking for `NaN`
values internally. We have also removed the reset function, as it was unused. If needed
in the future, it can be re-added.