refactor: use field initializers for injectees (#2258)
In this commit, we remove constructor parameters to ensure compatibility with the
`useDefineForClassFields` compiler option. When public class fields are enabled, such cases
throw a `TS2729: Property used before its initialization` error.
This occurs when a property (e.g., an injectee) is referenced before the constructor is executed.

Additionally, we are removing property decorators such as `@Inject` with this change.
Property decorators are not permitted when experimental decorators are disabled.
This adjustment helps make the framework compatible with these TypeScript compiler
options moving forward.