perf(store): prevent initializing state factory at feature levels
In this commit, we mark the state factory as a root provider to prevent its initialization
at feature levels when `provideStates` is called.

Previously, the state factory was initialized at feature levels because it needed to retrieve
the feature-level injector in order to call `injector.get(FeatureStateClass)`.
This was necessary because, when `NgxsModule.forFeature([BlogState])` was called, the `BlogState`
instance could only be retrieved from the feature injector hierarchy.

Now, since `addAndReturnDefaults` is called at the feature level inside the feature environment
initializer, it is executed within the feature injection context, allowing us to use the
`inject(...)` function directly.