refactor(store): make execution stategies tree-shakable (#2316)
NGXS has moved towards reducing bundle size by removing unnecessary code in production.

Currently, execution strategies are not tree-shakable because they are explicitly
referenced in the code. We should invert control and require developers to explicitly
provide an execution strategy.

When zoneless change detection is enabled and `NoopNgxsExecutionStrategy` is used,
`DispatchOutsideZoneNgxsExecutionStrategy` remains in the initial bundle, increasing
the final size even though it is never used — and vice versa.

With this commit, we make `options` a required parameter and `options.executionStrategy` a
required property, ensuring that developers always specify an execution strategy.

**BREAKING CHANGE**: The `options` parameter is now required when calling `forRoot()` or
`provideStore()`, as developers must provide an execution strategy. This should be a
straightforward change, as developers only need to specify
`{ executionStrategy: DispatchOutsideZoneNgxsExecutionStrategy }`.