feat(store): expose `AbortSignal` on the state context
This commit exposes the abort signal on the state context in action handlers.
It gives us a handle to proactively end async work. If we use an observable, it is
unsubscribed when cancellation happens. However, if we use `async/await`, there's no
way for the internal promise to know about the cancellation. As a result, the
`ctx.abortSignal.aborted` condition can be used to determine whether to proceed
after the `await` block.